test: fix import error
